extract languageService out of javascript mode
