+++ server/src/modes/script/javascript.ts
@@ -3,16 +3,14 @@ import { SymbolInformation, SymbolKind, CompletionItem, Location, SignatureHelp,
 import { LanguageMode } from '../languageModes';
 import { getWordAtText } from '../../utils/strings';
 import { VueDocumentRegions } from '../embeddedSupport';
-import { createUpdater, parseVue, isVue, getFileFsPath, getFilePath } from './preprocess';
+import { getFileFsPath, getFilePath } from './preprocess';
+import { getLanguageServie } from './languageService';
 
-import Uri from 'vscode-uri';
-import * as path from 'path';
 import * as ts from 'typescript';
 import * as _ from 'lodash';
 
 import { NULL_SIGNATURE, NULL_COMPLETION } from '../nullMode';
 
-import * as bridge from './bridge';
 
 const JS_WORD_REGEX = /(-?\d*\.\d\w*)|([^\`\~\!\@\#\%\^\&\*\(\)\-\=\+\[\{\]\}\\\|\;\:\'\"\,\.\<\>\/\?\s]+)/g;
 
@@ -29,143 +27,7 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
     return vueDocument.getEmbeddedDocument('javascript');
   });
 
-  let compilerOptions: ts.CompilerOptions = {
-    allowNonTsExtensions: true,
-    allowJs: true,
-    lib: ['lib.dom.d.ts', 'lib.es2017.d.ts'],
-    target: ts.ScriptTarget.Latest,
-    moduleResolution: ts.ModuleResolutionKind.NodeJs,
-    module: ts.ModuleKind.CommonJS,
-    allowSyntheticDefaultImports: true
-  };
-  let currentScriptDoc: TextDocument;
-  let versions = new Map<string, number>();
-  let scriptDocs = new Map<string, TextDocument>();
-
-  // Patch typescript functions to insert `import Vue from 'vue'` and `new Vue` around export default.
-  // NOTE: Typescript 2.3 should add an API to allow this, and then this code should use that API.
-  const { createLanguageServiceSourceFile, updateLanguageServiceSourceFile } = createUpdater();
-  (ts as any).createLanguageServiceSourceFile = createLanguageServiceSourceFile;
-  (ts as any).updateLanguageServiceSourceFile = updateLanguageServiceSourceFile;
-  const configFilename = ts.findConfigFile(workspacePath, ts.sys.fileExists, 'tsconfig.json') ||
-    ts.findConfigFile(workspacePath, ts.sys.fileExists, 'jsconfig.json');
-  const configJson = configFilename && ts.readConfigFile(configFilename, ts.sys.readFile).config || {
-    exclude: ['node_modules', '**/node_modules/*']
-  };
-  const parsedConfig = ts.parseJsonConfigFileContent(configJson,
-    ts.sys,
-    workspacePath,
-    compilerOptions,
-    configFilename,
-    undefined,
-    [{ extension: 'vue', isMixedContent: true }]);
-  const files = parsedConfig.fileNames;
-  compilerOptions = parsedConfig.options;
-  compilerOptions.allowNonTsExtensions = true;
-
-  function updateCurrentTextDocument (doc: TextDocument) {
-    const fileFsPath = getFileFsPath(doc.uri);
-    const filePath = getFilePath(doc.uri);
-    // When file is not in language service, add it
-    if (!scriptDocs.has(fileFsPath)) {
-      if (_.endsWith(fileFsPath, '.vue')) {
-        files.push(filePath);
-      }
-    }
-    if (!currentScriptDoc || doc.uri !== currentScriptDoc.uri || doc.version !== currentScriptDoc.version) {
-      currentScriptDoc = jsDocuments.get(doc);
-      let lastDoc = scriptDocs.get(fileFsPath);
-      if (lastDoc && currentScriptDoc.languageId !== lastDoc.languageId) {
-        // if languageId changed, restart the language service; it can't handle file type changes
-        jsLanguageService.dispose();
-        jsLanguageService = ts.createLanguageService(host);
-      }
-      scriptDocs.set(fileFsPath, currentScriptDoc);
-      versions.set(fileFsPath, (versions.get(fileFsPath) || 0) + 1);
-    }
-  }
-
-  const host: ts.LanguageServiceHost = {
-    getCompilationSettings: () => compilerOptions,
-    getScriptFileNames: () => files,
-    getScriptVersion (fileName) {
-      if (fileName === bridge.fileName) {
-        return '0';
-      }
-      const normalizedFileFsPath = getNormalizedFileFsPath(fileName);
-      let version = versions.get(normalizedFileFsPath);
-      return version ? version.toString() : '0';
-    },
-    getScriptKind (fileName) {
-      if (isVue(fileName)) {
-        const uri = Uri.file(fileName);
-        fileName = uri.fsPath;
-        const doc = scriptDocs.get(fileName) ||
-          jsDocuments.get(TextDocument.create(uri.toString(), 'vue', 0, ts.sys.readFile(fileName)));
-        return doc.languageId === 'typescript' ? ts.ScriptKind.TS : ts.ScriptKind.JS;
-      }
-      else {
-        if (fileName === bridge.fileName) {
-          return ts.Extension.Ts;
-        }
-        // NOTE: Typescript 2.3 should export getScriptKindFromFileName. Then this cast should be removed.
-        return (ts as any).getScriptKindFromFileName(fileName);
-      }
-    },
-    resolveModuleNames (moduleNames: string[], containingFile: string): ts.ResolvedModule[] {
-      // in the normal case, delegate to ts.resolveModuleName
-      // in the relative-imported.vue case, manually build a resolved filename
-      return moduleNames.map(name => {
-        if (name === bridge.moduleName) {
-          return {
-            resolvedFileName: bridge.fileName,
-            extension: ts.Extension.Ts
-          };
-        }
-        if (path.isAbsolute(name) || !isVue(name)) {
-          return ts.resolveModuleName(name, containingFile, compilerOptions, ts.sys).resolvedModule!;
-        }
-        const uri = Uri.file(path.join(path.dirname(containingFile), name));
-        const resolvedFileName = uri.fsPath;
-        if (ts.sys.fileExists(resolvedFileName)) {
-          const doc = scriptDocs.get(resolvedFileName) ||
-            jsDocuments.get(TextDocument.create(uri.toString(), 'vue', 0, ts.sys.readFile(resolvedFileName)));
-          return {
-            resolvedFileName,
-            extension: doc.languageId === 'typescript' ? ts.Extension.Ts : ts.Extension.Js,
-          };
-        }
-        return undefined as any;
-      });
-    },
-    getScriptSnapshot: (fileName: string) => {
-      if (fileName === bridge.fileName) {
-        let text = bridge.content;
-        return {
-          getText: (start, end) => text.substring(start, end),
-          getLength: () => text.length,
-          getChangeRange: () => void 0
-        };
-      }
-      const normalizedFileFsPath = getNormalizedFileFsPath(fileName);
-      let doc = scriptDocs.get(normalizedFileFsPath);
-      let text = doc ? doc.getText() : (ts.sys.readFile(normalizedFileFsPath) || '');
-      if (!doc && isVue(fileName)) {
-        // Note: This is required in addition to the parsing in embeddedSupport because
-        // this works for .vue files that aren't even loaded by VS Code yet.
-        text = parseVue(text);
-      }
-      return {
-        getText: (start, end) => text.substring(start, end),
-        getLength: () => text.length,
-        getChangeRange: () => void 0
-      };
-    },
-    getCurrentDirectory: () => workspacePath,
-    getDefaultLibFileName: ts.getDefaultLibFilePath,
-  };
-
-  let jsLanguageService = ts.createLanguageService(host);
+  let { jsLanguageService, updateCurrentTextDocument } = getLanguageServie(workspacePath, jsDocuments);
   let settings: any = {};
 
   return {
@@ -178,7 +40,7 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
       }
     },
     doValidation (doc: TextDocument): Diagnostic[] {
-      updateCurrentTextDocument(doc);
+      let scriptDoc = updateCurrentTextDocument(doc);
       if (!languageServiceIncludesFile(jsLanguageService, doc.uri)) {
         return [];
       }
@@ -191,26 +53,26 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
         // syntactic/semantic diagnostic always has start and length
         // so we can safely cast diag to TextSpan
         return {
-          range: convertRange(currentScriptDoc, diag as ts.TextSpan),
+          range: convertRange(scriptDoc, diag as ts.TextSpan),
           severity: DiagnosticSeverity.Error,
           message: ts.flattenDiagnosticMessageText(diag.messageText, '\n')
         };
       });
     },
     doComplete (doc: TextDocument, position: Position): CompletionList {
-      updateCurrentTextDocument(doc);
+      let scriptDoc = updateCurrentTextDocument(doc);
       if (!languageServiceIncludesFile(jsLanguageService, doc.uri)) {
         return { isIncomplete: false, items: [] };
       }
 
       const fileFsPath = getFileFsPath(doc.uri);
-      const offset = currentScriptDoc.offsetAt(position);
+      const offset = scriptDoc.offsetAt(position);
       const completions = jsLanguageService.getCompletionsAtPosition(fileFsPath, offset);
       if (!completions) {
         return { isIncomplete: false, items: [] };
       }
-      const wordAtText = getWordAtText(currentScriptDoc.getText(), offset, JS_WORD_REGEX);
-      const replaceRange = convertRange(currentScriptDoc, wordAtText);
+      const wordAtText = getWordAtText(scriptDoc.getText(), offset, JS_WORD_REGEX);
+      const replaceRange = convertRange(scriptDoc, wordAtText);
       const entries = completions.entries.filter(entry => entry.name !== '__vueEditorBridge');
       return {
         isIncomplete: false,
@@ -247,13 +109,13 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
       return item;
     },
     doHover (doc: TextDocument, position: Position): Hover {
-      updateCurrentTextDocument(doc);
+      let scriptDoc = updateCurrentTextDocument(doc);
       if (!languageServiceIncludesFile(jsLanguageService, doc.uri)) {
         return { contents: [] };
       }
 
       const fileFsPath = getFileFsPath(doc.uri);
-      const info = jsLanguageService.getQuickInfoAtPosition(fileFsPath, currentScriptDoc.offsetAt(position));
+      const info = jsLanguageService.getQuickInfoAtPosition(fileFsPath, scriptDoc.offsetAt(position));
       if (info) {
         const display = ts.displayPartsToString(info.displayParts);
         const doc = ts.displayPartsToString(info.documentation);
@@ -264,20 +126,20 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
           markedContents.unshift(doc, '\n');
         }
         return {
-          range: convertRange(currentScriptDoc, info.textSpan),
+          range: convertRange(scriptDoc, info.textSpan),
           contents: markedContents
         };
       }
       return { contents: [] };
     },
     doSignatureHelp (doc: TextDocument, position: Position): SignatureHelp {
-      updateCurrentTextDocument(doc);
+      let scriptDoc = updateCurrentTextDocument(doc);
       if (!languageServiceIncludesFile(jsLanguageService, doc.uri)) {
         return NULL_SIGNATURE;
       }
 
       const fileFsPath = getFileFsPath(doc.uri);
-      const signHelp = jsLanguageService.getSignatureHelpItems(fileFsPath, currentScriptDoc.offsetAt(position));
+      const signHelp = jsLanguageService.getSignatureHelpItems(fileFsPath, scriptDoc.offsetAt(position));
       if (!signHelp) {
         return NULL_SIGNATURE;
       }
@@ -313,17 +175,17 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
       return ret;
     },
     findDocumentHighlight (doc: TextDocument, position: Position): DocumentHighlight[] {
-      updateCurrentTextDocument(doc);
+      let scriptDoc = updateCurrentTextDocument(doc);
       if (!languageServiceIncludesFile(jsLanguageService, doc.uri)) {
         return [];
       }
 
       const fileFsPath = getFileFsPath(doc.uri);
-      const occurrences = jsLanguageService.getOccurrencesAtPosition(fileFsPath, currentScriptDoc.offsetAt(position));
+      const occurrences = jsLanguageService.getOccurrencesAtPosition(fileFsPath, scriptDoc.offsetAt(position));
       if (occurrences) {
         return occurrences.map(entry => {
           return {
-            range: convertRange(currentScriptDoc, entry.textSpan),
+            range: convertRange(scriptDoc, entry.textSpan),
             kind: <DocumentHighlightKind>(entry.isWriteAccess ? DocumentHighlightKind.Write : DocumentHighlightKind.Text)
           };
         });
@@ -331,7 +193,7 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
       return [];
     },
     findDocumentSymbols (doc: TextDocument): SymbolInformation[] {
-      updateCurrentTextDocument(doc);
+      let scriptDoc = updateCurrentTextDocument(doc);
       if (!languageServiceIncludesFile(jsLanguageService, doc.uri)) {
         return [];
       }
@@ -349,7 +211,7 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
               kind: convertSymbolKind(item.kind),
               location: {
                 uri: doc.uri,
-                range: convertRange(currentScriptDoc, item.spans[0])
+                range: convertRange(scriptDoc, item.spans[0])
               },
               containerName: containerLabel
             };
@@ -372,56 +234,56 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
       return [];
     },
     findDefinition (doc: TextDocument, position: Position): Definition {
-      updateCurrentTextDocument(doc);
+      let scriptDoc = updateCurrentTextDocument(doc);
       if (!languageServiceIncludesFile(jsLanguageService, doc.uri)) {
         return [];
       }
 
       const fileFsPath = getFileFsPath(doc.uri);
-      const definition = jsLanguageService.getDefinitionAtPosition(fileFsPath, currentScriptDoc.offsetAt(position));
+      const definition = jsLanguageService.getDefinitionAtPosition(fileFsPath, scriptDoc.offsetAt(position));
       if (!definition) {
         return [];
       }
       return definition.map(d => {
         return {
           uri: doc.uri,
-          range: convertRange(currentScriptDoc, d.textSpan)
+          range: convertRange(scriptDoc, d.textSpan)
         };
       });
     },
     findReferences (doc: TextDocument, position: Position): Location[] {
-      updateCurrentTextDocument(doc);
+      let scriptDoc = updateCurrentTextDocument(doc);
       if (!languageServiceIncludesFile(jsLanguageService, doc.uri)) {
         return [];
       }
 
       const fileFsPath = getFileFsPath(doc.uri);
-      const references = jsLanguageService.getReferencesAtPosition(fileFsPath, currentScriptDoc.offsetAt(position));
+      const references = jsLanguageService.getReferencesAtPosition(fileFsPath, scriptDoc.offsetAt(position));
       if (references) {
         return references.map(d => {
           return {
             uri: doc.uri,
-            range: convertRange(currentScriptDoc, d.textSpan)
+            range: convertRange(scriptDoc, d.textSpan)
           };
         });
       }
       return [];
     },
     format (doc: TextDocument, range: Range, formatParams: FormattingOptions): TextEdit[] {
-      updateCurrentTextDocument(doc);
+      let scriptDoc = updateCurrentTextDocument(doc);
 
       const fileFsPath = getFileFsPath(doc.uri);
       const initialIndentLevel = formatParams.scriptInitialIndent ? 1 : 0;
       const formatSettings = convertOptions(formatParams, settings && settings.format, initialIndentLevel);
-      const start = currentScriptDoc.offsetAt(range.start);
-      let end = currentScriptDoc.offsetAt(range.end);
+      const start = scriptDoc.offsetAt(range.start);
+      let end = scriptDoc.offsetAt(range.end);
       const edits = jsLanguageService.getFormattingEditsForRange(fileFsPath, start, end, formatSettings);
       if (edits) {
         const result = [];
         for (let edit of edits) {
           if (edit.span.start >= start && edit.span.start + edit.span.length <= end) {
             result.push({
-              range: convertRange(currentScriptDoc, edit.span),
+              range: convertRange(scriptDoc, edit.span),
               newText: edit.newText
             });
           }
@@ -455,10 +317,6 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
 
 }
 
-
-function getNormalizedFileFsPath (fileName: string): string {
-  return Uri.file(fileName).fsPath;
-}
 
 function languageServiceIncludesFile (ls: ts.LanguageService, documentUri: string): boolean {
   const filePaths = ls.getProgram().getRootFileNames();
